Validator now part of the exported jar. Can also be run as a gradle task with "gradle runValidator"
